fechas cambiadas de tipo, usecases creados

+++ app/src/main/java/es/etologic/mahjongscoring2/app/new_game/NewGameFragment.java
@@ -148,7 +148,7 @@ public class NewGameFragment extends Fragment {
         addPlayerChip(player);
     }
 
-    private void startGame(int gameId) {
+    private void startGame(long gameId) {
         
     }
